+++ src/components/LayerAttribute/components/ScaleSelector/index.tsx
@@ -1,0 +1,8 @@
+import type { ReactFC } from '@formily/react';
+import { connect } from '@formily/react';
+import type { SelectProps } from 'antd';
+import AntdScaleSelector from './AntdScaleSelector';
+
+const ScaleSelector: ReactFC<SelectProps> = connect(AntdScaleSelector);
+
+export default ScaleSelector;
