+++ src/components/LayerAttribute/components/index.ts
@@ -1,6 +1,6 @@
 export { ColorPicker, FieldSelect, RibbonSelect } from '@antv/insight-component';
-export { default as ColorType } from './ColorType';
 export { default as FormCollapse } from './FormCollapse';
 export { default as Offset } from './Offset';
+export { default as ScaleSelector } from './ScaleSelector';
 export { default as Slider } from './Slider';
 export { default as SliderRange } from './SliderRange';
